# perf: Improve performance when rendering elements