~28% faster.

Before:
```
Blueprint::HTML 0.11.0 749.34k (  1.33µs) (± 0.89%)  2.69kB/op   8.22× slower
                   ECR   6.16M (162.43ns) (± 0.67%)  1.51kB/op        fastest
```

After:
```
Blueprint::HTML 0.11.0 952.75k (  1.05µs) (± 0.56%)  2.69kB/op   6.47× slower
                   ECR   6.16M (162.24ns) (± 0.83%)  1.51kB/op        fastest
```